pyproject.toml
[tool.poetry]
name = "commitcrafter"
version = "0.2.2"
description = ""
authors = ["mpruvot <marius.pruvot@outlook.fr>"]
readme = "README.md"

[tool.poetry.scripts]
commitcraft = "commitcrafter.cli:app"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.1.1"
typer = "^0.12.3"
ruff = "^0.4.0"
gitpython = "^3.1.43"
openai = "^1.23.1"
rich = "^13.7.1"
anthropic = "^0.40.0"
pyperclip = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
